Removes garbage in code and inserts comments for routines
